Удаление классов: игрока, сущности и физики. Сделан класс уровня и карты

+++ ogl2/Sprite.cpp
@@ -326,7 +326,7 @@ void Sprite::draw()
     glBufferSubData(GL_ARRAY_BUFFER, 0, vertices.size() * sizeof(GLfloat), &vertices[0]);
 
     glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, this->EBO);
-    glBufferSubData(GL_ELEMENT_ARRAY_BUFFER, 0, indices.size() * sizeof(GLfloat), &indices[0]);
+    glBufferSubData(GL_ELEMENT_ARRAY_BUFFER, 0, indices.size() * sizeof(GLint), &indices[0]);
 
     glBindBuffer(GL_ARRAY_BUFFER, 0);
